Relocate TestLogAppender() to /util and all tests use
